Added function pointers for functions specific to board
